refactor(core): split AI service caller into modules
